fix merge errors and merge logic (fix #30)
The problem is that tinfoilhat model's merge method, while merging
source_files, assumes that src_uri cannot vary if the recipe version,
revision and variant is the same, and uses it as part of a temporary id
for source_files to merge.

The problem is that tinfoilhat model's merge method, while merging
source_files, assumes that src_uri cannot vary if the recipe version,
revision and variant is the same, and uses it as part of a temporary id
for source_files to merge.

However, we just discovered src_uri can vary even if recipe's version,
revision and variant are the same: in this specific case, it happened
for some gnome packages, for which src_uri was changed from
https://ftp.gnome.org/pub/GNOME/sources/... to
https://download.gnome.org/sources/....

BTW, we can safely ignore src_uri diffs and keep only relpath and
checksum as "unique together" identifiers for source_files.

Moreover, even if it is not related to this bug, we need to change the
whole merging logic for source_files, because here we just need to merge
source_files tags; actually, at that stage there must not be any new or
different source_files between the old and new .tinfoilhat.json files
that we want to merge, because we are merging metadata concerning the
same source variant (which by definition must have the same set of
source files)

Signed-off-by: Alberto Pianon <pianon@array.eu>