check if tinfoilhat file already been added before
Currently the add command adds and merges whatever tinfoilhat file it receives
as input. However, there may be a case (especially during tests, which at this
stage may be very frequent) where a tinfoilhat file with the same content has
been already added in the past, but it is merged again and added to the history,
making the history crowded and confusing.

A check should be added on the file checksum, in order to avoid to add and merge
the same data many times. Checking the checksum (and not the actual json
content) is enough, because add command does not copy the input file but parses
it and then dumps it into a fresh file, so even if the formatting of the input
file is different (eg. prettyprint) the output is always the same.

Signed-off-by: Alberto Pianon <alberto@pianon.eu>