option to add package variants before harvesting data
The Dashboard show audit progress also based on existing package variants - eg.
if there are multiple package variants, but only the oldest one has been
reviewed in Fossology, the Dashboard regards also the newer one as reviewed in
the total file count, because in variants only some single files (patches etc.)
are usually changed/added, so it's just a matter of a reuse agent that needs to
be scheduled in order to have also the new variant fully reviewed.

However, by harvesting only latest project snapshot's data, previous variants
are not included, so the total audited file count provided by the Dashboard is
not reliable.

To fix this, we need to add all available variants to the session before running
"fossy" and "harvest" commands, and properly handle them in the harvest command
option that filters packages based on latest snapshot tag (older variants
generally belong to a previous snapshot and may be filtered out unless a proper
exception is added to the filter)